Ensure we send the CSRF token cookie on login.
Summary:
Before diff, a new user would be unable to perform any POST requests.

Added a login_success view to which we redirect right after login.
This approach also has the benefit that we can now redirect to the
page the user was on before logging in.

Test Plan: Cleared cookies, logged in, logging out now works

Reviewers: bmatican

Reviewed By: bmatican

Differential Revision: http://phab.code4fun.de/D10